docker: setup docker & install clang python bindings
We were having trouble getting clang and it's python bindings to work on
everyones computers. This commit tries to setup a shared development
environment with docker to avoid these kinds of issues.

This dockerfile sets up an ubuntu container with llvm, clang and
python3 installed, and all the python libraries listed in requirements.txt.

It also downloads the clang source and adds its python bindings folder
to the PYTHONPATH to make them accessible to all.

To get this enviornment running:
  1. Build the image: `docker build --tag uob-summer-project .`
  2. Start a container with that image:
     `docker run --interactive --tty --volume $PWD:/project
  --publish-all uob-summer-project`

By default this launches a jupyter notebook server. If you just want a
shell, append `/bin/bash` to the end of the second command, i.e.:

    `docker run --interactive --tty --publish-all --volume $PWD:/project
    uob-summer-project /bin/bash`

I stuck some standard Python dependencies into requirements.txt to start.