docker,readme: modify Dockerfile & README.md to work on mac
Networking works slightly differently on a mac, and I needed to follow
the instructions on this stack overflow question:

  https://stackoverflow.com/questions/38830610/access-jupyter-notebook-running-on-docker-container

Essentially, you just need to tell jupyter to listen on all ip's, via
`--ip 0.0.0.0`.

I've updated the Dockerfile to do this as part of the default command,
and updated the README.md as well.