data,juliet: one juliet data point per file w/ multiple per testcase
After generating the split juliet dataset (where we separated out the
good examples from the bad for each test case), we noticed that some
data points consisted solely of comments! oh no...

The problem is that some test cases consist of multiple files. There
is a primary file, and usually one file for "good" behaviour and one
file for bad behaviour.

When we were treating each file as a test case, this meant that
extracting the bad code from a bad only file didn't give us a good
test case, it gave us an empty one.

To fix this:
  - we create one datapoint per file, but share the testcase_ID to
  allow linking later
  - duplicate each file when we create the good and bad split, then do
  the filource code filtering as before
  - [TODO] make the preprocess_code module aware of this, so that it:
    1. Combines the files in a way that clang can meaningfully parse
    them.
    2. Detect the primary file/entrypoint so that we can tell clang
    where to root it's tree.