adjacency: remove 25% biggest matrices #29
It turns out that a single matrix of size 44401x44401, represented with 8 bytes per entry, is 15gb! So it's not computationally feasible for us to do any analysis on these data points.

We checked and it seems that 75% of the data set can be fit 614x614 matrices, which only take aroun 3.38mb per datapoint. Much better!

This commit removes all the matrices bigger than 614x614, and parameterises the matrix size throughout the notebook.