adjacency: modify matrix format to fit in keras'Conv2D layer #29
Keras Conv2D is written to work with images, so it assumes that each
element in its 2D input matrix is actually a list of
values (implicitly, these correspond to the channels of the image,
e.g. the RGB values).

And so the Keras Conv2D requires a 3d input.

Our edges only have one "channel": the value 1 if an edge exists and 0
otherwise. To get this to work in Keras we need to convert our 2D
matrix into a "3D" one, like:

    / 0 0 1 \      / [0] [0] [1] \
    | 1 0 0 |  =>  | [1] [0] [0] |
    \ 0 1 0 /      \ [0] [1] [0] /

I've tried to do this with the wrap_elements function but it
fails. Any ideas?