adjacency: debug "kernel crash" problem + develop reasonable network #29
The "kernel crash" problem seems to have been caused by trying to allocate too much memory for each batch. We fixed this by reducing the batch size and running in google colab (more RAM) so that one batch is less than 10% of RAM.

We then developed a slightly more advanced network design and are currently letting it run.